requirements.txt
flask==2.2.2
python-dotenv==0.19.0
lark-parser==0.11.3
cffi==1.15.1
cryptography==3.4.7
openai>=1.0.0
requests>=2.26.0
pytest>=7.0.0
watchdog>=2.1.9
pyopenssl